Drop sa-awl output from pmg-system-report
After gathering some initial experience with `pmg-system-report` it seems
that the output of `sa-awl` is not too helpful in narrowing down problems.

Additionally the output tends to be quite large (multiple MiB) on larger/
longer-running installations, leading to timeouts or problems when sending
the report to the support.

Should the AWL checks skew the SA-results in a negative way this already shows
in the logs (which log the score for each rule).

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>