adapt postgresql.conf template to variable version
pass the postgres major version in template_vars->{postgres}->{version} and
use that in the template. Additionally use the version for the config file
location.

The start and end tags are changed to '<*' and '*>' since the original
postgresql.conf contains an occurence of the default start tag '[%' which
would need to get escaped

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>