Fix #2224: override reference to openspf.org
This is a fix for a small cosmetic issue, which keeps coming up on our various
support-channels:
pmgpolicy as user of 'Mail::SPF::Server' logs lines referencing the (now
defunct) url: http://www.openspf.org/Why

By overriding 'default_authority_explanation' [0] the logs now contain a simple
logmessage without any reference to an external url.

All macros used before are used in the new message as well, so the information
content stays the same.

Given that we are (by far) not the only users of this module I additionally
opened an issue upstream [1] (and linked it to the debian bug-report [2]).

Tested by providing a SPF policy not allowing the ip of the sender in my
testsetup and observing the logs.

[0] https://metacpan.org/pod/Mail::SPF::Server
[1] https://rt.cpan.org/Public/Bug/Display.html?id=130413
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=930032

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>