add support for sending NDRs on Block
This patch adds a flag to the mail-section of pmg.conf - 'ndr_on_block',
defaulting to false.

If enabled pmg-smtp-filter replies with 554 (permanent fail) for the recipient
who blocked the mail.

Since pmg-smtp-filter is currently called with LMTP it can respond with a
separate code for each recipient. Postfix records the answers and generates
and sends the NDR to the sender.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>