api: statistics: add common method for details
This patch changes adds a new method, which yields the detail statistics
for a particular address.

Depending on the provided type argument it returns the same information as
the contact/sender/receiver detail calls.

This allows the statistics to be displayed for addresses containing
'/' in their localpart, once this is permitted in our api schema.

the idea follows a similar change for the user blocklists in
e8d909c11faeb5a4f84f39ef50e0eaf8ea65046d

By adding a new API method we can eventually drop the old methods with 7.0

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>