config: fix "var declared in conditional statement"
This is actually buggy and can lead to unexpected issues as in the
case the check on the declared variable did not evaluates to true it
gets (or better keeps) the value from the previous time when it was
actually assigned. Found with perlcritic, which reports the highest
severity for this mistake.

Refactor out the "is current file equal to generated config" check
which fixes three instances of that on its own and reduces code bloat
a bit.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>