fix #3915: remove obsolete /etc/apt/apt.conf.d/75pmgconf
this was forgotten when introducing the more flexible kernel-keeping
logic with proxmox-boot-tool (in 6.4).

with this file present no pve-kernel gets autoremoved.

this patch uses d/maintscript for removing instead of using
debian/conffiles (deb-conffiles(5)) 'remove-on-upgrade'

sticking with d/maintscript was chosen, since else it depends on the
installed debhelper version if the removal is done at all (debhelper
from bullseye simply ignores remove-on-upgrade in d/conffiles)

Tested the following with a local version bump to 7.1-5 and a VM:
* regular unchanged /etc/apt/apt.conf.d/75pmgconf
* manually modified /etc/apt/apt.conf.d/75pmgconf
* manually removed /etc/apt/apt.conf.d/75pmgconf

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>